refactor(sidebar): extract sidebar component and conversation loader refactor(sidebar): サイドバーと会話読み込みサービスを分離 - Extract sidebar UI rendering into a dedicated component - Move conversation data fetching and mapping logic to a separate service - Wire the extracted modules in app.py without changing existing behaviour - サイドバーのUIレンダリングを専用コンポーネントとして分離しました - 会話データの取得とマッピングのロジックを独立したサービスへ移行しました - 既存の挙動を変更せずに、抽出したモジュールをapp.py内で結合
